kdam
Struct Bar

pub struct Bar {

    pub animation: Animation,
    pub colour: Option<Colour>,
    pub desc: String,
    pub delay: f32,
    pub disable: bool,
    pub dynamic_miniters: bool,
    pub dynamic_ncols: bool,
    pub force_refresh: bool,
    pub inverse_unit: bool,
    pub leave: bool,
    pub mininterval: f32,
    pub miniters: usize,
    pub ncols: Option<u16>,
    pub position: u16,
    pub postfix: String,
    pub total: usize,
    pub unit: String,
    pub unit_divisor: usize,
    pub unit_scale: bool,
    pub writer: Writer,
    pub bar_length: u16,
    pub counter: usize,
    /* private fields */

}

Core implemention of console progress bar.

Example

use kdam::{tqdm, Bar, BarExt};

let mut pb = Bar::new(100);
// let mut pb = tqdm!(total = 100);
// let mut pb = Bar::builder().total(100).build().unwrap();

for _ in 0..100 {
    pb.update(1).unwrap();
}

Fields
animation: Animationcolour: Option<Colour>desc: Stringdelay: f32disable: booldynamic_miniters: booldynamic_ncols: boolforce_refresh: boolinverse_unit: boolleave: boolmininterval: f32miniters: usizencols: Option<u16>position: u16postfix: Stringtotal: usizeunit: Stringunit_divisor: usizeunit_scale: boolwriter: Writerbar_length: u16counter: usize
Implementations

impl Bar

pub fn new(total: usize) -> Self

Create a new Bar with a total value.

Example

use kdam::Bar;

let pb = Bar::new(100);

pub fn builder() -> BarBuilder

Create a new BarBuilder.

Example

use kdam::Bar;

let pb = Bar::builder().total(100).build().unwrap();

pub fn set_description<T: Into<String>>(&mut self, description: T)

Set/Modify description property.

pub fn set_postfix<T: Into<String>>(&mut self, postfix: T)

Set/Modify postfix property.

pub fn completed(&self) -> bool

Returns whether progress is completed or not.

If total is 0, it always returns false.

pub fn elapsed_time(&mut self) -> f32

Set and returns progress elapsed time.

pub fn fmt_counter(&self) -> String

Returns formatted counter value.

pub fn fmt_elapsed_time(&self) -> String

Returns formatted elapsed time.

pub fn fmt_percentage(&self, precision: usize) -> String

Returns formatted progress percentage.

pub fn fmt_rate(&self) -> String

Returns formatted progress rate.

pub fn fmt_remaining_time(&self) -> String

Returns formatted remaining time.

pub fn fmt_total(&self) -> String

Returns formatted total value.

pub fn indefinite(&self) -> bool

Returns whether progress is indefinite i.e. total is 0 or not.

pub fn ncols_for_animation(&mut self, padding: u16) -> u16

Set and returns number of columns for bar animation with given padding.

pub fn percentage(&self) -> f32

Returns progress percentage, like 0.62, 0.262, 1.0.

If total is 0, it always returns 1.0.

pub fn rate(&self) -> f32

Returns progress/iterations rate.

Note

Before calling this method, elapsed_time method should be called.

pub fn remaining_time(&self) -> f32

Returns remaining time (ETA) for progress completion.

If total is 0, it always returns infinity.

pub fn should_refresh(&mut self) -> bool

Returns whether to trigger a display update or not.

pub fn started(&self) -> bool

Returns whether progress is started i.e. counter is 0 or not.