kdam
Module term

Terminal related utilities.

Enums

InitializedOutput

Initialized output for writer.

Writer

Stderr and Stdout writer.

Traits

Colorizer

Add ANSI colour escape codes to the given text for printing coloured text in terminal.

Functions

colour

Create ANSI colour escape code from primary colours, hex code, rgb(r,g,b) and ansi(n).

hide_cursor

Hide cursor.

init

Enable/Disable colorization property of colorizer trait.

show_cursor

Show cursor.

width

Get terminal width.